+++ services/tok/init.go
@@ -31,6 +31,7 @@ func Init() {
 	drupal.CheckSettings()
 	docker.FindOrCreateTokCompose()
 	ssh.GenerateKeys()
+	conf.SetDrupalConfig()
 	git.IgnoreDefaults()
 
 	// Run Unison for syncing